==> data_preprocessing/1_GOSlim_add_domain.ipynb <==
# Assign GOSlim domain to each GOSlim accession

## Background

When obtained from Ensembl Biomart, GOSlim data does not have domain information(e.g., cellular_component) listed, so use QuickGO to assign domains to each GOSlim accession.

import requests
import time
import os
#import pandas as pd
import polars as pl #Replace with pandas

## Get domain function through QuickGO API

def get_domain(accession):
    requestURL = f"https://www.ebi.ac.uk/QuickGO/services/ontology/go/search?query={accession}&limit=1&page=1"
    try:
        r = requests.get(requestURL, headers={"Accept": "application/json"}, timeout=30)
        r.raise_for_status()
    except requests.exceptions.RequestException as e:
        print(f"API request error: {e}")
        return None
    data = r.json()
    return data['results'][0]['aspect'] if data['results'] else None

## MOUSE

ensembl_goslim_mouse = pl.read_csv("../data/biomart_goslim/biomart_mouse_goslim_R110.tsv", # Ensembl release 110
                                   separator="\t",
                                   infer_schema_length=0) # Convert all columns to string

#ensembl_goslim_mouse_unique = ensembl_goslim_mouse.unique(subset=["Gene stable ID", "GOSlim GOA Accession(s)","GOSlim GOA Description"], maintain_order=True)
ensembl_goslim_mouse_unique = ensembl_goslim_mouse.unique(keep="first", maintain_order=True)
display(ensembl_goslim_mouse_unique)

unique_goslim_terms_mouse = ensembl_goslim_mouse_unique.unique(subset=['GOSlim GOA Accession(s)']).select('GOSlim GOA Accession(s)')
display(unique_goslim_terms_mouse)

# pandas -> mouse_goslim_df["GOSlim_domain"] = mouse_goslim_df["GOSlim GOA Accession(s)"].apply(get_domain)
# https://docs.pola.rs/py-polars/html/reference/series/api/polars.Series.map_elements.html
mouse_goslim_df = unique_goslim_terms_mouse.with_columns(
    unique_goslim_terms_mouse["GOSlim GOA Accession(s)"].map_elements(get_domain, return_dtype=str).alias("GOSlim_domain")
)
time.sleep(10)

display(mouse_goslim_df)

# pandas -> mouse_merged_goslim_domain = pd.merge(ensembl_goslim_mouse_unique, mouse_goslim_df, how='left', on='GOSlim GOA Accession(s)')
mouse_merged_goslim_domain = ensembl_goslim_mouse_unique.join(
    mouse_goslim_df, 
    on="GOSlim GOA Accession(s)", 
    how="left")
display(mouse_merged_goslim_domain)

## HUMAN

ensembl_goslim_human = pl.read_csv("../data/biomart_goslim/biomart_human_goslim_R110.tsv", 
                                   separator="\t",
                                   infer_schema_length=0) # Ensembl release 110
ensembl_goslim_human_unique = ensembl_goslim_human.unique(keep="first", maintain_order=True)
display(ensembl_goslim_human_unique)

unique_goslim_terms_human = ensembl_goslim_human_unique.unique(subset=['GOSlim GOA Accession(s)']).select('GOSlim GOA Accession(s)')
display(unique_goslim_terms_human)

# pandas -> human_goslim_df["GOSlim_domain"] = human_goslim_df["GOSlim GOA Accession(s)"].apply(get_domain)
human_goslim_df = unique_goslim_terms_human.with_columns(
    unique_goslim_terms_human["GOSlim GOA Accession(s)"].map_elements(get_domain, return_dtype=str).alias("GOSlim_domain")
)
time.sleep(10)
display(human_goslim_df)

# pandas -> human_merged_goslim_domain = pd.merge(ensembl_goslim_human_unique, human_goslim_df,how="left",on="GOSlim GOA Accession(s)")

human_merged_goslim_domain = ensembl_goslim_human_unique.join(
    human_goslim_df, 
    on="GOSlim GOA Accession(s)", 
    how="left")
display(human_merged_goslim_domain)

## RICE

ensembl_goslim_rice = pl.read_csv("../data/biomart_goslim/biomart_rice_goslim_R56.tsv", 
                                  separator="\t",
                                  infer_schema_length=0) # Ensembl release 56
ensembl_goslim_rice_unique = ensembl_goslim_rice.unique(keep="first", maintain_order=True)
display(ensembl_goslim_rice_unique)

unique_goslim_terms_rice = ensembl_goslim_rice_unique.unique(subset=['GOSlim GOA Accession(s)']).select('GOSlim GOA Accession(s)')
display(unique_goslim_terms_rice)

rice_goslim_df = unique_goslim_terms_rice.with_columns(
    unique_goslim_terms_rice["GOSlim GOA Accession(s)"].map_elements(get_domain, return_dtype=str).alias("GOSlim_domain")
)
time.sleep(10)

display(rice_goslim_df)

rice_merged_goslim_domain = ensembl_goslim_rice_unique.join(
    rice_goslim_df, 
    on="GOSlim GOA Accession(s)", 
    how="left"
)

display(rice_merged_goslim_domain)

## Save the results

mouse_merged_goslim_domain.write_csv("../data/biomart_goslim/biomart_mouse_goslim_R110_domain.tsv", separator="\t")
human_merged_goslim_domain.write_csv("../data/biomart_goslim/biomart_human_goslim_R110_domain.tsv", separator="\t")
rice_merged_goslim_domain.write_csv("../data/biomart_goslim/biomart_rice_goslim_R56_domain.tsv", separator="\t")
