==> data_preprocessing/2_GOSlim_combined_all_HR.ipynb <==
## GOSlim combined all genes based on the GOslim terms (version 0.5)

#import pandas as pd
#import dask.dataframe as dd
import polars as pl

organism1_vs_organism2 = 'HR' # human vs rice

goslim_all_human = pl.read_csv("../data/biomart_goslim/biomart_human_goslim_R110_domain.tsv", 
                               separator="\t",
                               infer_schema_length=0,
                               low_memory=False)
goslim_all_rice = pl.read_csv("../data/biomart_goslim/biomart_rice_goslim_R56_domain.tsv", 
                              separator="\t",
                              infer_schema_length=0,
                              low_memory=False)
columns_of_interest = ["Gene stable ID", "GOSlim GOA Accession(s)", "GOSlim GOA Description", "GOSlim_domain"]

goslim_all_human_filtered = goslim_all_human[columns_of_interest]
goslim_all_rice_filtered = goslim_all_rice[columns_of_interest]

display(goslim_all_human_filtered)
display(goslim_all_rice_filtered)

## Merge dataframe based on common column

#merge dataframe based on common column 'GOSlim GOA Accession(s)', 'GOSlim GOA Description', 'GOSlim_domain'
merged_df_goslim = goslim_all_human_filtered.join(
    goslim_all_rice_filtered,
    left_on=['GOSlim GOA Accession(s)', 'GOSlim GOA Description', 'GOSlim_domain'],
    right_on=['GOSlim GOA Accession(s)','GOSlim GOA Description', 'GOSlim_domain'],
    how='inner',
    suffix='_rice'
)
merged_df_goslim = merged_df_goslim.rename({"Gene stable ID" : "Gene stable ID_human"})
merged_df_goslim = merged_df_goslim.select([
    "GOSlim GOA Accession(s)",
    "GOSlim GOA Description",
    "GOSlim_domain",
    "Gene stable ID_human",
    "Gene stable ID_rice"
])

display(merged_df_goslim)

### Count the number of common columns

counted_df = merged_df_goslim.group_by(["GOSlim GOA Accession(s)", "GOSlim GOA Description"]).count()
sorted_df = counted_df.sort("count", descending=True)
display(sorted_df)

### Debugging

#create set of unique accessions
unique_accessions_h = set(goslim_all_human_filtered['GOSlim GOA Accession(s)'].unique())
unique_accessions_r = set(goslim_all_rice_filtered['GOSlim GOA Accession(s)'].unique())

#check if there are common accessions between human and rice
common_accessions = unique_accessions_r.intersection(unique_accessions_h)
unique_accessions_merged = set(merged_df_goslim["GOSlim GOA Accession(s)"].unique())
missing_accessions = common_accessions.difference(unique_accessions_merged)

#check duplicates
duplicates = merged_df_goslim.is_duplicated() #pandas -> 'duplicated' function
unique_combinations = not duplicates.any() #check if there are any duplicates(return boolean)

#output
print(f"Number of unique accessions in human: {len(unique_accessions_h)}")
print(f"Number of unique accessions in rice: {len(unique_accessions_r)}")
print(f"Number of common accessions: {len(common_accessions)}")

if missing_accessions:
    print(f"Number of missing accessions: {len(missing_accessions)}")
    print(f"Missing accessions: {missing_accessions}")
else:
    print("Number of missing accessions: 0")

print(f"Are the combinations of 'Gene stable ID_rice' and 'Gene stable ID_human' unique?: {unique_combinations}")

#write output
merged_df_goslim.write_csv(f"../data/GOSlim_merge_{organism1_vs_organism2}/GOslim_merge_common_goslim_correspondence_all.tsv", 
                           separator="\t")

### Extract dataset based on GOslim domain (BP, CC, MF)

# if you search specific string in column, use 'filter' function and 'str.contains' method
cellular_component = merged_df_goslim.filter(pl.col("GOSlim_domain").str.contains("cellular_component"))
molecular_function = merged_df_goslim.filter(pl.col("GOSlim_domain").str.contains("molecular_function"))
biological_process = merged_df_goslim.filter(pl.col("GOSlim_domain").str.contains("biological_process"))


#write output
cellular_component.write_csv(f"../data/GOSlim_merge_{organism1_vs_organism2}/GOslim_merge_common_goslim_correspondence_cc.tsv",
                            separator="\t")
molecular_function.write_csv(f"../data/GOSlim_merge_{organism1_vs_organism2}/GOslim_merge_common_goslim_correspondence_mf.tsv",
                            separator="\t")
biological_process.write_csv(f"../data/GOSlim_merge_{organism1_vs_organism2}/GOslim_merge_common_goslim_correspondence_bp.tsv",
                            separator="\t")


display(cellular_component, molecular_function, biological_process)
