==> data_preprocessing/3_GOSlim_uniq_up.ipynb <==
import  pandas as pd

## HUMAN

extract_df_h_goslim = pd.read_csv('../data/human_annotation/human_annotation_up.tsv', sep='\t')
uniq_term_h = extract_df_h_goslim['GOSlim GOA Accession(s)'].unique()
df_uniq_term_h = pd.DataFrame(uniq_term_h, columns=['Unique GOSlim GOA Accession(s)'])

# get descriptions from extract_df_h_goslim
descriptions = []
for term in uniq_term_h:
    description = extract_df_h_goslim[extract_df_h_goslim['GOSlim GOA Accession(s)'] == term]['GOSlim GOA Description'].iloc[0]
    descriptions.append(description)

df_uniq_term_h['GOSlim GOA Description'] = descriptions

df_uniq_term_h.to_csv('../data/human_annotation/human_annotation_up_uniq.tsv', sep='\t', index=False)

display(extract_df_h_goslim, df_uniq_term_h)

## RICE

extract_df_r_goslim = pd.read_csv('../data/rice_annotation/rice_annotation_up.tsv', sep='\t')

uniq_term_r = extract_df_r_goslim['GOSlim GOA Accession(s)'].unique()
df_uniq_term_r = pd.DataFrame(uniq_term_r, columns=['Unique GOSlim GOA Accession(s)'])

# get descriptions from extract_df_r_goslim
descriptions = []
for term in uniq_term_r:
    description = extract_df_r_goslim[extract_df_r_goslim['GOSlim GOA Accession(s)'] == term]['GOSlim GOA Description'].iloc[0]
    descriptions.append(description)

df_uniq_term_r['GOSlim GOA Description'] = descriptions
df_uniq_term_r.to_csv('../data/rice_annotation/rice_annotation_up_uniq.tsv', sep='\t', index=False)

display(extract_df_r_goslim, df_uniq_term_r)

## COMMON TERMS (HUMAN & RICE)

common_term = pd.merge(df_uniq_term_h, 
                      df_uniq_term_r, 
                      on=['Unique GOSlim GOA Accession(s)', 'GOSlim GOA Description'],
                      how='inner')

common_term.to_csv('../data/circos_up_HR_2311/common_term_up.tsv', sep='\t', index=False)
display(common_term)

## MOUSE

extract_df_m_goslim = pd.read_csv('../data/mouse_annotation/mouse_annotation_up.tsv', sep='\t')
uniq_term_m = extract_df_m_goslim['GOSlim GOA Accession(s)'].unique()
df_uniq_term_m = pd.DataFrame(uniq_term_m, columns=['Unique GOSlim GOA Accession(s)'])

# get descriptions from extract_df_h_goslim
descriptions = []
for term in uniq_term_m:
    description = extract_df_m_goslim[extract_df_m_goslim['GOSlim GOA Accession(s)'] == term]['GOSlim GOA Description'].iloc[0]
    descriptions.append(description)

df_uniq_term_m['GOSlim GOA Description'] = descriptions

df_uniq_term_m.to_csv('../data/mouse_annotation/mouse_annotation_up_uniq.tsv', sep='\t', index=False)

display(extract_df_m_goslim, df_uniq_term_m)

## COMMON TERMS (MOUSE & RICE)

common_term_m = pd.merge(df_uniq_term_m, 
                      df_uniq_term_r, 
                      on=['Unique GOSlim GOA Accession(s)', 'GOSlim GOA Description'],
                      how='inner')

common_term_m.to_csv('../data/circos_up_MR_2312/common_term_up.tsv', sep='\t', index=False)
display(common_term_m)
