==> data_preprocessing/5_combined_df_human.ipynb <==
import pandas as pd
import os

# output directory and settings
organism = 'human'
annotation_dir = f'../data/{organism}_annotation'
os.makedirs(annotation_dir, exist_ok=True)

# human HN-score data
# https://doi.org/10.6084/m9.figshare.23944935.v2
df = pd.read_csv(f"../data/HN-score_{organism}/HN-score_{organism}_all.csv", sep=",")


# Upregulated genes
df_up = pd.read_csv(f"../data/HN-score_{organism}/HN5_genes_up_{organism}.tsv", sep="\t") #genes names in this file were manually corrected to the most recent version 
df_up['GeneName'] = df_up['GeneName'].str.upper()
# Downregulated genes
df_down = pd.read_csv(f"../data/HN-score_{organism}/HN5_genes_down_{organism}.tsv", sep="\t")
df_down['GeneName'] = df_down['GeneName'].str.upper()

display(df, df_up, df_down)

## Convert the gene names to Ensemble gene IDs

df2 = pd.read_csv("../data/gprofiler/gProfiler_hsapiens_2023-11-13 10-00-00.csv", sep=",")
df2_ID = df2[["initial_alias", "converted_alias", "name"]]

display(df2_ID)

# Get ensemble gene ID from gProfiler UP 1%

# check duplicates
duplicates = df2_ID[df2_ID.duplicated(subset=["initial_alias"], keep=False)]

# upregulated genes
merged_df_up = pd.merge(  df_up, 
                          df2_ID, 
                          left_on="GeneName", 
                          right_on="initial_alias", 
                          how="inner")

gene_names_up = set(df_up['GeneName'])
gene_names_merged = set(merged_df_up['GeneName'])
missing_gene_names = gene_names_up - gene_names_merged
missing_rows_up = df_up[df_up['GeneName'].isin(missing_gene_names)]

UP = merged_df_up[["converted_alias", "HN5", "name"]].copy()
UP.rename(columns={'HN5': 'HN-score(HN5)'}, inplace=True)

# downregulated genes
merged_df_down = pd.merge(df_down,
                           df2_ID,
                            left_on="GeneName",
                            right_on="initial_alias",
                            how="inner")


gene_names_down = set(df_down['GeneName'])
gene_names_merged_down = set(merged_df_down['GeneName'])
missing_gene_names_down = gene_names_down - gene_names_merged_down
missing_rows_down = df_down[df_down['GeneName'].isin(missing_gene_names_down)]

DOWN = merged_df_down[["converted_alias", "HN5", "name"]].copy()
DOWN.rename(columns={'HN5': 'HN-score(HN5)'}, inplace=True)

COMBINED = pd.concat(
    [UP, DOWN],
    ignore_index=True
)

# if missing_rows_up and missing_rows_down are empty, print message
if missing_rows_up.empty and missing_rows_down.empty:
    print("All gene names were successfully converted to ensemble gene IDs.")
else:
    print("Some gene names were not converted to ensemble gene IDs.")
    print("Missing gene names in UP:")
    display(missing_rows_up)
    print("Missing gene names in DOWN:")
    display(missing_rows_down)


display(UP, DOWN, COMBINED)

## merge the Ensembl biomart data (Chromosome position)

- For drawing Circos plot

df_ensembl_position = pd.read_csv(f"../data/biomart_goslim/biomart_{organism}_position_R110.tsv", sep="\t", low_memory=False)

# 1. merged all data
merged_df_position = pd.merge(
    df_ensembl_position,
    COMBINED,
    left_on="Gene stable ID",
    right_on="converted_alias",
    how="inner"
)
merged_df_position = merged_df_position.drop(columns=['converted_alias'])
merged_df_position = merged_df_position.sort_values(by='HN-score(HN5)', ascending=False).copy()


# 2. merged up genes data
merged_df_position_up = pd.merge(
    df_ensembl_position,
    UP,
    left_on="Gene stable ID",
    right_on="converted_alias",
    how="inner"
)
merged_df_position_up = merged_df_position_up.drop(columns=['converted_alias'])
merged_df_position_up = merged_df_position_up.sort_values(by='HN-score(HN5)', ascending=False).copy()


# 3. merged down genes data
merged_df_position_down = pd.merge(
    df_ensembl_position,
    DOWN,
    left_on="Gene stable ID",
    right_on="converted_alias",
    how="inner"
)
merged_df_position_down = merged_df_position_down.drop(columns=['converted_alias'])
merged_df_position_down = merged_df_position_down.sort_values(by='HN-score(HN5)', ascending=False).copy()

# rename columns
chromosome_position = [ merged_df_position, merged_df_position_up, merged_df_position_down ]

for df in chromosome_position:
    df.rename(columns={'Gene start (bp)': 'start2 (bp)'}, inplace=True)
    df.rename(columns={'Gene end (bp)': 'end2 (bp)'}, inplace=True)

# save data
merged_df_position.to_csv(f"../data/{annotation_dir}/{organism}_position_all.tsv", sep="\t" , index=False)
merged_df_position_up.to_csv(f"../data/{annotation_dir}/{organism}_position_up.tsv", sep="\t", index=False)
merged_df_position_down.to_csv(f"../data/{annotation_dir}/{organism}_position_down.tsv", sep="\t", index=False)

display(df_ensembl_position, merged_df_position, merged_df_position_up, merged_df_position_down)

## merge the Ensembl biomart data (GOslim)

- For exexution of GO enrichment analysis

df_ensembl = pd.read_csv(f"../data/biomart_goslim/biomart_{organism}_goslim_R110_domain.tsv", sep="\t", low_memory=False)
df_ensembl_uniq = df_ensembl.drop_duplicates(subset=['Gene stable ID', 
                                                     'GOSlim GOA Accession(s)', 
                                                     'GOSlim GOA Description'], 
                                                     keep='first').copy()
df_ensembl_uniq.rename(columns={'start2 (bp)': 'Gene start(bp)'}, inplace=True)
df_ensembl_uniq.rename(columns={'end2 (bp)': 'Gene end(bp)'}, inplace=True)

# 1. merge all data
merged_df = pd.merge(df_ensembl_uniq,COMBINED,
                        left_on="Gene stable ID",
                        right_on="converted_alias",
                     how="inner")

merged_df = merged_df.drop(columns=['converted_alias'])
merged_df = merged_df.sort_values(by='HN-score(HN5)', ascending=False).copy()
merged_df.to_csv(f"../data/{annotation_dir}/{organism}_annotation.tsv", sep="\t" , index=False)

# 2. merge UP data
merged_df_up = pd.merge(df_ensembl_uniq,UP,
                        left_on="Gene stable ID",
                        right_on="converted_alias",
                     how="inner")

merged_df_up = merged_df_up.drop(columns=['converted_alias'])
merged_df_up = merged_df_up.sort_values(by='HN-score(HN5)', ascending=False).copy()


# 3. merge DOWN data
merged_df_down = pd.merge(df_ensembl_uniq,DOWN,
                        left_on="Gene stable ID",
                        right_on="converted_alias",
                     how="inner")

merged_df_down = merged_df_down.drop(columns=['converted_alias'])
merged_df_down = merged_df_down.sort_values(by='HN-score(HN5)', ascending=False).copy()


# save data
merged_df.to_csv(f"../data/{annotation_dir}/{organism}_annotation.tsv", sep="\t" , index=False)
merged_df_up.to_csv(f"../data/{annotation_dir}/{organism}_annotation_up.tsv", sep="\t" , index=False)
merged_df_down.to_csv(f"../data/{annotation_dir}/{organism}_annotation_down.tsv", sep="\t", index=False)

display(df_ensembl_uniq, merged_df, merged_df_up, merged_df_down)
